fix(client): Fix Chromium error with file upload

+++ client/src/utils/ui.rs
@@ -4,7 +4,7 @@ pub(crate) fn format_date_time(naive_date_time: NaiveDateTime) -> String {
     naive_date_time.format("%d/%m/%Y, %H:%M:%S").to_string()
 }
 
-pub(crate) fn format_number_as_ordinal(number: u32) -> String {
+pub(crate) fn format_number_as_ordinal(number: usize) -> String {
     match number {
         1 => "first".to_string(),
         2 => "second".to_string(),
